requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
neo4j==5.19.0
python-multipart==0.0.12
ultralytics>=8.3.50
pydantic==2.10.0
python-dotenv==1.0.0
Pillow>=10.3.0
numpy>=2.0.0
